Clearing issues with merge and code
